Revert "Remove function "Administrator Access""
This reverts commit 1854f82e34bb58b2946120485d041f197c20c9ef.

+++ FlightManagementSystem/src/sample/controller/FlightList.java
@@ -38,6 +38,9 @@ public class FlightList extends ArrayList<Flight> implements I_FlightList {
     List<Reservation> reservationList = new ArrayList<>();
     List<BoardingPass> boardingPassList = new ArrayList<>();
     List<CrewAssignment> crewAssignmentList = new ArrayList<>();
+    User admin = new User("admin@123", true);
+
+    boolean loginCheck = false;
     String fileName = "Product.dat";
 
 //------------------------------------------------------------------------------
@@ -51,7 +54,11 @@ public class FlightList extends ArrayList<Flight> implements I_FlightList {
 
     @Override
     public void flightSchelduleManagement() {
-        add();
+        if (loginCheck) {
+            add();
+        } else if (!loginCheck) {
+            System.out.println("You have no authority to access this function.");
+        }
     }
 
     public void add() {
@@ -93,12 +100,17 @@ public class FlightList extends ArrayList<Flight> implements I_FlightList {
 //==============================================================================
     @Override
     public void passengerReservationAndBooking() {
-        List<Flight> availableFlights = findAvailableFlights();
-        if (!availableFlights.isEmpty()) {
-            for (Flight f : availableFlights) {
-                System.out.println(f);
+        if (!loginCheck) {
+            List<Flight> availableFlights = findAvailableFlights();
+            if (!availableFlights.isEmpty()) {
+                for(Flight f:availableFlights)
+                {
+                    System.out.println(f);
+                }
+                makeReservation(availableFlights);
             }
-            makeReservation(availableFlights);
+        } else if (loginCheck) {
+            System.out.println("You have no authority to access this function.");
         }
     }
 
@@ -151,7 +163,11 @@ public class FlightList extends ArrayList<Flight> implements I_FlightList {
 //==============================================================================
     @Override
     public void passengerCheckInAndSeatAllocation() {
-        passengerCheckInFlights();
+        if (!loginCheck) {
+            passengerCheckInFlights();
+        } else if (loginCheck) {
+            System.out.println("You have no authority to access this function.");
+        }
     }
 
     private void passengerCheckInFlights() {
@@ -231,7 +247,11 @@ public class FlightList extends ArrayList<Flight> implements I_FlightList {
 //==============================================================================
     @Override
     public void crewManagementAndAssignments() {
-        addCrewAssignments();
+        if (loginCheck) {
+            addCrewAssignments();
+        } else if (!loginCheck) {
+            System.out.println("You have no authority to access this function.");
+        }
     }
 
     public void addCrewAssignments() {
@@ -254,37 +274,68 @@ public class FlightList extends ArrayList<Flight> implements I_FlightList {
         } while (!hasFound);
 
     }
-
 //==============================================================================
+
+    @Override
+    public void administratorAccessForSystemManagement() {
+        login();
+    }
+
+    private void login() {
+        boolean check = false;
+        String loginChoice = Utils.getString("Do you want to login as Admin/User (A/U): ");
+        if (loginChoice.toUpperCase().equals("A")) {
+            do {
+                String logPassWord = Utils.getString("Input password: ");
+                if (logPassWord.equals(admin.getPassword()) && admin.isAdmin()) {
+                    loginCheck = true;
+                    System.out.println("Login with adminstator successfully !!!");
+                } else if (!logPassWord.equals(admin.getPassword())) {
+                    System.out.println("Incorrect password !!!");
+                    check = Utils.confirmYesNo("Do you want to continue login (Y/N): ");
+                }
+            } while (check);
+
+        } else if (loginChoice.toUpperCase().equals("U")) {
+            System.out.println("Login with user successfully !!!");
+            loginCheck = false;
+        }
+    }
+//==============================================================================
+
     @Override
     public void save() {
-        File file = new File(fileName);
-        if (file.exists()) {
+        if (loginCheck) {
+            File file = new File(fileName);
+            if (file.exists()) {
 
-        } else {
+            } else {
+                try {
+                    file.createNewFile();
+                } catch (IOException ex) {
+                    Logger.getLogger(FlightList.class.getName()).log(Level.SEVERE, null, ex);
+                }
+            }
             try {
-                file.createNewFile();
+                FileOutputStream fos = new FileOutputStream(file);
+                ObjectOutputStream oos = new ObjectOutputStream(fos);
+
+                oos.writeObject(this);
+                oos.writeObject(reservationList);
+                oos.writeObject(crewAssignmentList);
+
+                System.out.println("Save to file successfully !!!");
+            } catch (FileNotFoundException ex) {
+                Logger.getLogger(FlightList.class.getName()).log(Level.SEVERE, null, ex);
             } catch (IOException ex) {
                 Logger.getLogger(FlightList.class.getName()).log(Level.SEVERE, null, ex);
             }
+        } else if (!loginCheck) {
+            System.out.println("You have no authority to access this function.");
         }
-        try {
-            FileOutputStream fos = new FileOutputStream(file);
-            ObjectOutputStream oos = new ObjectOutputStream(fos);
+    }
 
-            oos.writeObject(this);
-            oos.writeObject(reservationList);
-            oos.writeObject(crewAssignmentList);
-
-            System.out.println("Save to file successfully !!!");
-        } catch (FileNotFoundException ex) {
-            Logger.getLogger(FlightList.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (IOException ex) {
-            Logger.getLogger(FlightList.class.getName()).log(Level.SEVERE, null, ex);
-        }
-}
-
-public void load() throws IOException {
+    public void load() throws IOException {
         File file = new File(fileName);
         if (file.exists()) {
